+++ src/manifolds/Rotations.jl
@@ -408,7 +408,8 @@ and that means the inverse has to be appliead to the (Euclidean) Hessian
 to map it into the Lie algebra.
 """
 riemannian_Hessian(M::Rotations, p, G, H, X)
-function riemannian_Hessian!(::Rotations{N}, Y, p, G, H, X) where {N}
+function riemannian_Hessian!(M::Rotations, Y, p, G, H, X)
+    N = get_n(M)
     symmetrize!(Y, G' * p)
     project!(SkewSymmetricMatrices(N), Y, p' * H - X * Y)
     return Y